Exporting abi, networks and bytecode as separate JSON files.
Changed package-npm.js script to take either a string with a file path
or an object as the value for the source files map.

If an object is provided, each key will represent a key in the JSON
source file that you want to save in a separate output file, and its
value will be the destination file.